# cmd/api/docs/swagger.yaml
basePath: /
definitions:
...
  model.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  model.Project:
    properties:
      active:
        type: boolean
      created_at:
        type: string
      deleted_at:
        type: string
      description:
        type: string
      link:
        type: string
      name:
        type: string
      profile_id:
        type: integer
      project_id:
        type: string
      updated_at:
        type: string
    type: object
  model.Resume:
    properties:
      certification_ids:
        description: IDs das certificações selecionadas
        items:
          type: string
        type: array
      contact_ids:
        description: IDs dos contatos selecionados
        items:
          type: string
        type: array
      education_ids:
        description: IDs das formações selecionadas
        items:
          type: string
        type: array
      experience_ids:
        description: IDs das experiências selecionadas
        items:
          type: string
        type: array
      id:
        type: string
      language_ids:
        description: IDs dos idiomas selecionados
        items:
          type: string
        type: array
      profile_id:
        type: integer
      project_ids:
        description: IDs dos projetos selecionados
        items:
          type: string
        type: array
      skill_ids:
        description: IDs das skills selecionadas
        items:
          type: string
        type: array
      social_media_ids:
        description: IDs das redes sociais selecionadas
        items:
          type: string
        type: array
    type: object
  model.SkillDTO:
    properties:
      embeddings:
        items:
          type: number
        type: array
      level:
        type: string
      name:
        type: string
      profile_id:
        type: integer
      skill_id:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: API para gerenciamento de perfis e currículos com IA.
  title: Resume GenAI API
  version: "1.0"
paths:
...
  /projects/:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Project'
            type: array
      summary: Lista todos os projetos
      tags:
      - projects
    post:
      consumes:
      - application/json
      parameters:
      - description: Projeto a ser criado
        in: body
        name: project
        required: true
        schema:
          $ref: '#/definitions/model.Project'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Project'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Cria um novo projeto
      tags:
      - projects
  /projects/{id}:
    delete:
      parameters:
      - description: ID do projeto
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Remove um projeto
      tags:
      - projects
    get:
      parameters:
      - description: ID do projeto
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Project'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Busca projeto por ID
      tags:
      - projects
    put:
      consumes:
      - application/json
      parameters:
      - description: ID do projeto
        in: path
        name: id
        required: true
        type: string
      - description: Projeto atualizado
        in: body
        name: project
        required: true
        schema:
          $ref: '#/definitions/model.Project'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Project'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Atualiza um projeto
      tags:
      - projects
  /resumes/:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Resume'
            type: array
      summary: Lista todos os currículos
      tags:
      - resumes
    post:
      consumes:
      - application/json
      parameters:
      - description: Currículo a ser criado
        in: body
        name: resume
        required: true
        schema:
          $ref: '#/definitions/model.Resume'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Resume'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Cria um novo currículo
      tags:
      - resumes
  /resumes/{id}:
    delete:
      parameters:
      - description: ID do currículo
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Remove um currículo
      tags:
      - resumes
    get:
      parameters:
      - description: ID do currículo
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Resume'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Busca currículo por ID
      tags:
      - resumes
    put:
      consumes:
      - application/json
      parameters:
      - description: ID do currículo
        in: path
        name: id
        required: true
        type: string
      - description: Currículo atualizado
        in: body
        name: resume
        required: true
        schema:
          $ref: '#/definitions/model.Resume'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Resume'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Atualiza um currículo
      tags:
      - resumes
  /skills/:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.SkillDTO'
            type: array
      summary: Lista todas as skills
      tags:
      - skills
    post:
      consumes:
      - application/json
      parameters:
      - description: Skill a ser criada
        in: body
        name: skill
        required: true
        schema:
          $ref: '#/definitions/model.SkillDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.SkillDTO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Cria uma nova skill
      tags:
      - skills
  /skills/{id}:
    delete:
      parameters:
      - description: ID da skill
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Remove uma skill
      tags:
      - skills
    get:
      parameters:
      - description: ID da skill
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.SkillDTO'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Busca skill por ID
      tags:
      - skills
    put:
      consumes:
      - application/json
      parameters:
      - description: ID da skill
        in: path
        name: id
        required: true
        type: string
      - description: Skill atualizada
        in: body
        name: skill
        required: true
        schema:
          $ref: '#/definitions/model.SkillDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.SkillDTO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Atualiza uma skill
      tags:
      - skills
...
swagger: "2.0"